bot.py

import logging
from telegram import Update, Bot
from telegram.ext import ApplicationBuilder, CommandHandler, MessageHandler, filters, ContextTypes
import requests
import os

TOKEN = "8131148603:AAFzVpTJrQeOkYU8qd74lh-ToPKLDUYYsgk"
REPLICATE_API_TOKEN = "r8_Pik**********************************"

# Configure logging
logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("Welcome to Photo Magic AI! Send a photo and use /enhance, /ghibli, or /bgremove.")

async def handle_photo(update: Update, context: ContextTypes.DEFAULT_TYPE):
    photo_file = await update.message.photo[-1].get_file()
    file_path = f"{update.message.from_user.id}_photo.jpg"
    await photo_file.download_to_drive(file_path)
    context.user_data["photo"] = file_path
    await update.message.reply_text("Photo received! Now send /enhance or /ghibli or /bgremove")

async def enhance(update: Update, context: ContextTypes.DEFAULT_TYPE):
    if "photo" not in context.user_data:
        await update.message.reply_text("Please send a photo first.")
        return

    photo_path = context.user_data["photo"]
    await update.message.reply_text("Enhancing your photo...")

    url = "https://api.replicate.com/v1/predictions"
    headers = {
        "Authorization": f"Token {REPLICATE_API_TOKEN}",
        "Content-Type": "application/json"
    }
    data = {
        "version": "92858f8f27054f8c845d4e49b1c7795d4c7e5d370beff8fdfacbff576f1af0ee",
        "input": {
            "img": open(photo_path, "rb").read()
        }
    }

    await update.message.reply_text("Feature coming soon in full version!")

async def ghibli(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("Ghibli Art generation is under development.")

async def bgremove(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("Background remover feature coming soon.")

def main():
    app = ApplicationBuilder().token(TOKEN).build()
    app.add_handler(CommandHandler("start", start))
    app.add_handler(CommandHandler("enhance", enhance))
    app.add_handler(CommandHandler("ghibli", ghibli))
    app.add_handler(CommandHandler("bgremove", bgremove))
    app.add_handler(MessageHandler(filters.PHOTO, handle_photo))
    app.run_polling()

if __name__ == "__main__":
    main()
